// src/generated/api/iam/iam.schemas.ts
export type AssignUserRoleDtoScopeType =
  (typeof AssignUserRoleDtoScopeType)[keyof typeof AssignUserRoleDtoScopeType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AssignUserRoleDtoScopeType = {
  BRANCH: 'BRANCH',
} as const;

export interface AssignUserRoleDto {
  roleCode: AssignableStaffRoleCode;
  scopeType: AssignUserRoleDtoScopeType;
  /** @pattern ^[1-9][0-9]*$ */
  branchId: string;
}

export type UserRoleAssignmentDtoStatus =
  (typeof UserRoleAssignmentDtoStatus)[keyof typeof UserRoleAssignmentDtoStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserRoleAssignmentDtoStatus = {
  ACTIVE: 'ACTIVE',
} as const;

export interface UserRoleAssignmentDto {
  /** @pattern ^[1-9][0-9]*$ */
  id: string;
  /** @pattern ^[1-9][0-9]*$ */
  userId: string;
  /** @pattern ^[1-9][0-9]*$ */
  roleId: string;
  roleCode: string;
  scopeType: ScopeType;
  /** @pattern ^[1-9][0-9]*$ */
  branchId?: string;
  status: UserRoleAssignmentDtoStatus;
  validFrom: string;
}
// ...
export type AssignableStaffRoleCode =
  (typeof AssignableStaffRoleCode)[keyof typeof AssignableStaffRoleCode];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AssignableStaffRoleCode = {
  BRANCH_MANAGER: 'BRANCH_MANAGER',
  STAFF: 'STAFF',
} as const;

export type ScopeType = (typeof ScopeType)[keyof typeof ScopeType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ScopeType = {
  GLOBAL: 'GLOBAL',
  BRANCH: 'BRANCH',
} as const;